5.1.21 Перенесет логику отрисовки карточки/формы в контроллер
