+++ src/controllers/board.js
@@ -1,2 +1,38 @@
+import {render, RenderPosition, replace} from "../utils/render";
+import TaskComponent from "../components/task";
+import TaskEditComponent from "../components/task-edit";
+
+const renderTask = (taskListElement, task) => {
+  const onEscKeyDown = (evt) => {
+    const isEscKey = evt.key === `Escape` || evt.key === `Esc`;
+
+    if (isEscKey) {
+      replaceEditToTask();
+      document.removeEventListener(`keydown`, onEscKeyDown);
+    }
+  };
+
+  const replaceEditToTask = () => {
+    replace(taskComponent, taskEditComponent);
+  };
+
+  const replaceTaskToEdit = () => {
+    replace(taskEditComponent, taskComponent);
+  };
+
+  const taskComponent = new TaskComponent(task);
+
+  taskComponent.setEditButtonClickHandler(() => {
+    replaceTaskToEdit();
+    document.addEventListener(`keydown`, onEscKeyDown);
+  });
+
+  const taskEditComponent = new TaskEditComponent(task);
+
+  taskEditComponent.setSubmitHandler(replaceEditToTask);
+
+  render(taskListElement, taskComponent, RenderPosition.BEFOREEND);
+};
+
 export default class BoardController {
 }
